The plugin now disables during relevent situations
The whole plugin will disable itself if a non AS3 project is opened. That
way, if you press Test and you're not dealing with an AS3 project then the
spaceport plugin won't hijack "TEST" access. I still need to make it so if
Spaceport plugin fails to load correctly then it will disable as well.