Fixed watcher reporting broken before it starte
This was caused by the fact that it only had a "Working State" and a "Non
working state". Before it even tried to start it was in a "Non working
state". I fixed this by changing the state to an enum which can have three
states. Started, FailedToStart, and NotStarted.